+++ test/ClientTest.php
@@ -2,10 +2,9 @@
 declare(strict_types=1);
 namespace FreeMobile;
 
-use function PHPUnit\Expect\{await, expect, fail, it};
+use function PHPUnit\Expect\{expect, fail, it};
 use PHPUnit\Framework\{TestCase};
 use Psr\Http\Message\{UriInterface};
-use Rx\Subject\{Subject};
 
 /**
  * Tests the features of the `FreeMobile\Client` class.
@@ -26,48 +25,42 @@ class ClientTest extends TestCase {
   }
 
   /**
-   * @test Client::onRequest
-   */
-  public function testOnRequest() {
-    it('should return an `Observable` instead of the underlying `Subject`', function() {
-      expect((new Client)->onRequest())->to->not->be->instanceOf(Subject::class);
-    });
-  }
-
-  /**
-   * @test Client::onResponse
-   */
-  public function testOnResponse() {
-    it('should return an `Observable` instead of the underlying `Subject`', function() {
-      expect((new Client)->onResponse())->to->not->be->instanceOf(Subject::class);
-    });
-  }
-
-  /**
    * @test Client::sendMessage
    */
   public function testSendMessage() {
-    it('should not send valid messages with invalid credentials', await(function() {
-      (new Client('', ''))->sendMessage('Hello World!')->subscribe(
-        function() { fail('A message with empty credentials should not be sent.'); },
-        function() { expect(true)->to->be->true; }
-      );
-    }));
+    it('should not send valid messages with invalid credentials', function() {
+      try {
+        (new Client('', ''))->sendMessage('Hello World!');
+        fail('A message with empty credentials should not be sent.');
+      }
 
-    it('should not send invalid messages with valid credentials', await(function() {
-      (new Client('anonymous', 'secret'))->sendMessage('')->subscribe(
-        function() { fail('A message with empty credentials should not be sent.'); },
-        function() { expect(true)->to->be->true; }
-      );
-    }));
+      catch (\Throwable $e) {
+        expect(true)->to->be->true;
+      }
+    });
+
+    it('should not send invalid messages with valid credentials', function() {
+      try {
+        (new Client('anonymous', 'secret'))->sendMessage('');
+        fail('A message with empty credentials should not be sent.');
+      }
+
+      catch (\Throwable $e) {
+        expect(true)->to->be->true;
+      }
+    });
 
     if (is_string($username = getenv('FREEMOBILE_USERNAME')) && is_string($password = getenv('FREEMOBILE_PASSWORD'))) {
-      it('should send valid messages with valid credentials', await(function() use ($password, $username) {
-        (new Client($username, $password))->sendMessage('Bonjour Cédric !')->subscribe(
-          function() { expect(true)->to->be->true; },
-          function(\Throwable $e) { fail($e->getMessage()); }
-        );
-      }));
+      it('should send valid messages with valid credentials', function() use ($password, $username) {
+        try {
+          (new Client($username, $password))->sendMessage('Bonjour Cédric !');
+          expect(true)->to->be->true;
+        }
+
+        catch (\Throwable $e) {
+          fail($e->getMessage());
+        }
+      });
     }
   }